240522 00:21 taehun profileImgView-add

+++ final_buk04_11_05/src/main/java/com/ssafy/enjoytrip/user/service/UserServiceImpl.java
@@ -42,7 +42,9 @@ public class UserServiceImpl implements UserService{
 		String saveFolder = projectPath  + File.separator + "final_05" + File.separator + "src" + File.separator + "assets" + File.separator + "users";
 		String filePath = saveFolder + File.separator + userInfo.getImg();
 		File file = new File(filePath);
-		file.delete();
+		if (!userDto.getImg().equals(userInfo.getImg())) {
+			file.delete();
+		}
 		return userDao.modify(userDto);
 	}
